+++ src/config/defaults.ts
@@ -1,0 +1,7 @@
+import { Viewport } from "@/types/viewport";
+
+export const defaultViewport: Viewport = {
+  longitude: -118.482505,
+  latitude: 34.0248477,
+  zoom: 14,
+};
