+++ src/types/viewport.ts
@@ -1,0 +1,5 @@
+export interface Viewport {
+  longitude: number;
+  latitude: number;
+  zoom: number;
+}
